Extract out path resolution from checksum cli command.
It's useful to be able to resolve a URL or an absolute/relative path.  Use case: Stack().read("http://xyz") or Stack().read("ISS/fov_001/experiment.json")

Renames resolve_url to a better name and improve docblock.  Note that this will break earlier versions of starfish that expect resolve_url to exist.