Make cache concurrency safe.
If someone evicts our cache entry between the time we check the cache and the time we read the cache, we get a KeyError.  See https://travis-ci.org/spacetx/starfish/builds/417492246 for an example of this happening.

Now we try to return the cache entry, and if that fails, we try to fetch the data.  When we fetch the data, we return it directly from the buffer instead of the cache to prevent the race condition from happening again.