Fix the checksum calculation in the cli command.
The job of tile writing is delegated to two components: a opener, which returns a context manager that acts like a file, and a writer, which takes that file-like object and writes the data to it.

The checksum cli command works by creating a non-operational context manager, paired with a writer that does nothing.  Unfortunately, to calculate the checksum, we need to actually write to the file.

This fix changes the context manager to have a write method that does nothing, and a writer that writes the original content of the file[1].  This way, we don't write the tile files.

[1] If a tile writer is not provided, then by default, it writes the tile data as a numpy array.