enum34: remove dependency on py34+
For platforms that do not need enum34, conda packages
are not being built. As a part of getting starfish
built in bioconda, the enum34 dependency has been
disabled as an install dependency. pkg_resources'
load_entry_point however still assumes that enum34
is required due to setup.py and fails with:

```
pkg_resources.DistributionNotFound:
   The 'enum34' distribution was not found
   and is required by slicedimage
```

see:
 - https://github.com/spacetx/starfish/issues/44
 - https://github.com/bioconda/bioconda-recipes/pull/10637