Always calculate checksums when writing tiles.
The checksums attached to the tiles are the checksums at the time the tile is loaded.  It should not be used to write the tiles.

This should address the issue in this [comment](https://github.com/spacetx/starfish/issues/458#issuecomment-415213814)

Test plan: Add a test that writes a TIFF tileset, loads it back up, writes as a NUMPY tileset, and load that back up.  Without the change in slicedimage/io.py, it writes the checksum for the TIFF file and the load (and the test) fails.  With the change, the test passes.