Add a numpy-array-future data source for tiles.
Currently, if we want to use a future to provide image data for a tile, we have to wrap it in a BytesIO object and then use `set_source_fh_contextmanager`.  This removes the necessity for wrapping.

We need to use this interface to load data into memory more selectively, rather than all at once.

Test plan: `make -j test`