Record the tile shape in our data structures if it's not provided.
If the tile_shape is not provided by the json file, we should infer it from the actual data.  This is a suboptimal data path since it requires a decoding step just to get the shape of the tile, so we warn about it.

Previously, this field was set by `_load()`, but that was removed in #72.

Test plan: Generate a tileset document, and then strip the shape data from the tiles.  Then read it back.  The correct tile shapes should be provided, but getting them should trigger a warning.