Allow partitions to be written to any relative path.
Currently, we only take the basename of the partition path generated by `partition_path_generator`.  We should accept any relative path.  This change removes the constraint that the partition be in the same path as the parent path, and adds a test to ensure that it works.

Partially addresses https://github.com/spacetx/starfish/issues/1116