Parse collection documents in parallel.
Adds a `parse_collection` method that binds the parse_method, baseurl, and backend_config.  That allows for a method that simply accepts the name + the relative path or url (as a tuple).

This method is mapped over a threadpool to yield an iterable of name, partition tuples.  In the main thread, we add the partitions to the collection and return it.