Add retry for HTTP failures and network connectivity issues
In the HTTP backend, we create a urllib3's retry policy and attach it to a python-requests session.

We test this by monkeypatching the code to retry on 404 errors (which we normally do not).  We attempt to fetch a file that's not present, but start a thread that creates the file after a short delay.  It should initially not find the file, and then eventually succeed.

Fixes https://github.com/spacetx/starfish/issues/1277