Fix conversion of file:// URLs to local paths.
Previously, we just extracted the path component and used that as the base path for the DiskCache backend.  However, on Windows, we need to convert the path separators to forward slashes.  We do this by parsing the URL into a posix path, and then creating a local system path from the path components.

Test plan: would be great if this was tested on Windows....