Clean up our usage of DiskCache
In our caching tests, we create a temporary directory, and then initialize a cache in that temporary directory.  The issue is that the cache handle holds an open file handle, and on Windows, the cleanup of the temporary directory fails if the cache handle still exists.

The fix is to add a context manager to remove the cache handle.

Test plan: works on windows
Part of https://github.com/spacetx/starfish/issues/1296