Close files before using them in tests
It's not possible, using python stdlib, to open a file in Windows for write and read simultaneously.  A common idiom in the test code is to set up a temporary file using `with tempfile.NamedTemporaryFile()`, write some content to it, and then switch to the test code, and then cleaning up the temp file when the `with` statement goes out of scope.

This is incompatible with Windows, so we instead just create the temporary file, write out the necessary contents, and then close it (and not delete it).  We rely on the temporary directory cleanup to ensure that the files eventually get cleaned up.

On the */test_disk_backend.py, we were not using a temporary directory, so we add that.

Test plan: works on windows
Part of https://github.com/spacetx/starfish/issues/1296